WIP - Add unit tests to the shopify refunder
